requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
python-multipart==0.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
openai==1.3.7
pandas==2.1.4
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
fastapi-users==12.1.3
httpx==0.25.2
aiofiles==23.2.1
python-magic==0.4.27
requests==2.31.0
jinja2==3.1.2
pydantic-settings==2.0.3
